Added default Jenkins project to run.
For use cases when you may not have a specific project to run for a repo, but a generic project to run for all repos. Configuring the same project for every repo could become tedious. Instead I added `default_project`, which will be used when no specific project is set for a repo. This allows a more flexible project configuration.